setup global cennznet wallet
