+++ .env.development
@@ -1,3 +1,3 @@
-NEXT_PUBLIC_SUPPORTED_ASSETS=1,2,1000,1001,1002
+NEXT_PUBLIC_SUPPORTED_ASSETS=16000,16001,17002
 NEXT_PUBLIC_API_URL="wss://nikau.centrality.me/public/ws"
 NEXT_PUBLIC_ETH_CHAIN_ID=42
