+++ components/shared/WalletModal.tsx
@@ -1,13 +1,9 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { Box, Button, CircularProgress, Modal } from "@mui/material";
-import {
-	StyledModal,
-	Heading,
-	SmallText,
-	Option,
-} from "../../theme/StyledComponents";
+import { StyledModal, Heading, Option } from "../../theme/StyledComponents";
 import { useWeb3Accounts } from "../../providers/Web3AccountsProvider";
 import { useWallet } from "../../providers/SupportedWalletProvider";
+import AccountBalances from "./AccountBalances";
 
 const WalletModal: React.FC<{
 	setModalOpen: Function;
@@ -15,89 +11,16 @@ const WalletModal: React.FC<{
 	modalState: string;
 }> = ({ setModalOpen, setModalState, modalState }) => {
 	const [open] = useState(true);
-	const { getBridgeBalances, bridgeBalances, selectedAccount, selectAccount } =
-		useWallet();
+	const { selectAccount, selectedAccount } = useWallet();
 	const accounts = useWeb3Accounts();
-
-	useEffect(() => {
-		if (selectedAccount && !bridgeBalances)
-			getBridgeBalances(selectedAccount.address);
-	}, [bridgeBalances, getBridgeBalances, selectedAccount]);
 
 	const updateAccount = (account) => {
 		if (account !== selectedAccount) {
 			selectAccount(account);
 			window.localStorage.setItem("CENNZnet-account", account);
-			getBridgeBalances(selectedAccount.address);
 		}
 		setModalState("showWallet");
 	};
-
-	const AccountBalances = selectedAccount && (
-		<>
-			<Box sx={{ mt: "5%", pl: "5%", display: "flex" }}>
-				<Heading
-					sx={{
-						color: "primary.main",
-						fontSize: "18px",
-
-						textTransform: "uppercase",
-					}}
-				>
-					{selectedAccount.meta.name}&nbsp;
-				</Heading>
-				<Heading
-					sx={{
-						color: "black",
-						fontSize: "18px",
-						display: "flex",
-					}}
-				>
-					{"[Account Name]"}
-				</Heading>
-			</Box>
-			<SmallText sx={{ pl: "5%", opacity: "70%" }}>
-				{selectedAccount.address}
-			</SmallText>
-			{bridgeBalances ? (
-				<Box sx={{ mt: "3%", pl: "5%", display: "block" }}>
-					{Object.values(bridgeBalances).map((token: any, i) => (
-						<Box key={i}>
-							<SmallText
-								sx={{
-									color: "black",
-									fontSize: "18px",
-									display: "inline-flex",
-								}}
-							>
-								{token.symbol} Balance:
-							</SmallText>
-							<SmallText
-								sx={{
-									color: "black",
-									fontWeight: "bold",
-									fontSize: "18px",
-									display: "inline-flex",
-								}}
-							>
-								{token.balance}
-							</SmallText>
-							<br />
-						</Box>
-					))}
-				</Box>
-			) : (
-				<>
-					<SmallText
-						sx={{ color: "primary.main", fontSize: "14", margin: "10px auto" }}
-					>
-						Fetching Balances...
-					</SmallText>
-					<CircularProgress sx={{ margin: "0 auto" }} />
-				</>
-			)}
-		</>
-	);
 
 	return (
 		<Modal open={open}>
@@ -128,7 +51,9 @@ const WalletModal: React.FC<{
 				</Box>
 				{modalState === "showWallet" ? (
 					<>
-						{AccountBalances}
+						{selectedAccount && (
+							<AccountBalances selectedAccount={selectedAccount} />
+						)}
 						<Button
 							sx={{
 								fontFamily: "Teko",
