+++ pages/exchange/index.tsx
@@ -3,9 +3,7 @@ import { Box, Button, TextField } from "@mui/material";
 import TokenPicker from "../../components/shared/TokenPicker";
 import ExchangeIcon from "../../components/exchange/ExchangeIcon";
 import { useCENNZApi } from "../../providers/CENNZApiProvider";
-import SupportedAssetsProvider, {
-	useAssets,
-} from "../../providers/SupportedAssetsProvider";
+import { useAssets } from "../../providers/SupportedAssetsProvider";
 import { Amount, AmountUnit } from "../../utils/exchange/Amount";
 import { Asset } from "../../types/exchange";
 import BigNumber from "bignumber.js";
@@ -55,116 +53,114 @@ const Exchange: React.FC<{}> = () => {
 	}, [api, exchangeTokenValue, assets, exchangeToken, receivedToken]);
 
 	return (
-		<SupportedAssetsProvider>
-			<Box
-				sx={{
+		<Box
+			sx={{
+				position: "absolute",
+				top: "25%",
+				left: "calc(50% - 552px/2)",
+				display: "flex",
+				justifyContent: "center",
+				alignItems: "center",
+			}}
+		>
+			<Button
+				style={{
 					position: "absolute",
-					top: "25%",
-					left: "calc(50% - 552px/2)",
+					top: "-5%",
+					width: "100%",
 					display: "flex",
 					justifyContent: "center",
 					alignItems: "center",
+					backgroundColor: "#FFFFFF",
+					border: "4px solid #1130FF",
+					flex: "none",
+					order: 0,
+					alignSelf: "stretch",
+					flexGrow: 1,
+					margin: "0px 0px",
+					borderBottom: "none",
+					fontFamily: "Teko",
+					fontWeight: "bold",
+					fontSize: "24px",
+					lineHeight: "124%",
+					color: "#1130FF",
 				}}
 			>
+				Swap Tokens
+			</Button>
+			<Box
+				component="form"
+				sx={{
+					width: "552px",
+					height: "auto",
+					margin: "0 auto",
+					background: "#FFFFFF",
+					border: "4px solid #1130FF",
+					display: "flex",
+					flexDirection: "column",
+					justifyContent: "center",
+					alignItems: "center",
+					padding: "0px",
+				}}
+			>
+				<TokenPicker
+					setToken={setExchangeToken}
+					cennznet={true}
+					forceSelection={exchangeToken}
+				/>
+				<TextField
+					label="Amount"
+					variant="outlined"
+					required
+					sx={{
+						width: "80%",
+						m: "30px 0 30px",
+					}}
+					value={exchangeTokenValue}
+					onChange={(event) => setExchangeTokenValue(event.target.value)}
+				/>
+				<ExchangeIcon
+					onClick={() => {
+						setReceivedToken(exchangeToken);
+						setExchangeToken(receivedToken);
+						setExchangeTokenValue(receivedTokenValue);
+						setReceivedTokenValue(exchangeTokenValue);
+					}}
+				/>
+				<TokenPicker
+					setToken={setReceivedToken}
+					cennznet={true}
+					forceSelection={receivedToken}
+					removeToken={exchangeToken}
+				/>
+				<TextField
+					label="Amount"
+					variant="outlined"
+					required
+					sx={{
+						width: "80%",
+						m: "30px 0 30px",
+					}}
+					value={receivedTokenValue}
+					onChange={(event) => setReceivedTokenValue(event.target.value)}
+				/>
 				<Button
-					style={{
-						position: "absolute",
-						top: "-5%",
-						width: "100%",
-						display: "flex",
-						justifyContent: "center",
-						alignItems: "center",
-						backgroundColor: "#FFFFFF",
-						border: "4px solid #1130FF",
-						flex: "none",
-						order: 0,
-						alignSelf: "stretch",
-						flexGrow: 1,
-						margin: "0px 0px",
-						borderBottom: "none",
+					sx={{
 						fontFamily: "Teko",
 						fontWeight: "bold",
-						fontSize: "24px",
+						fontSize: "21px",
 						lineHeight: "124%",
 						color: "#1130FF",
+						mt: "30px",
+						mb: "50px",
 					}}
+					size="large"
+					variant="outlined"
 				>
-					Swap Tokens
+					Exchange
 				</Button>
-				<Box
-					component="form"
-					sx={{
-						width: "552px",
-						height: "auto",
-						margin: "0 auto",
-						background: "#FFFFFF",
-						border: "4px solid #1130FF",
-						display: "flex",
-						flexDirection: "column",
-						justifyContent: "center",
-						alignItems: "center",
-						padding: "0px",
-					}}
-				>
-					<TokenPicker
-						setToken={setExchangeToken}
-						cennznet={true}
-						forceSelection={exchangeToken}
-					/>
-					<TextField
-						label="Amount"
-						variant="outlined"
-						required
-						sx={{
-							width: "80%",
-							m: "30px 0 30px",
-						}}
-						value={exchangeTokenValue}
-						onChange={(event) => setExchangeTokenValue(event.target.value)}
-					/>
-					<ExchangeIcon
-						onClick={() => {
-							setReceivedToken(exchangeToken);
-							setExchangeToken(receivedToken);
-							setExchangeTokenValue(receivedTokenValue);
-							setReceivedTokenValue(exchangeTokenValue);
-						}}
-					/>
-					<TokenPicker
-						setToken={setReceivedToken}
-						cennznet={true}
-						forceSelection={receivedToken}
-						removeToken={exchangeToken}
-					/>
-					<TextField
-						label="Amount"
-						variant="outlined"
-						required
-						sx={{
-							width: "80%",
-							m: "30px 0 30px",
-						}}
-						value={receivedTokenValue}
-						onChange={(event) => setReceivedTokenValue(event.target.value)}
-					/>
-					<Button
-						sx={{
-							fontFamily: "Teko",
-							fontWeight: "bold",
-							fontSize: "21px",
-							lineHeight: "124%",
-							color: "#1130FF",
-							mt: "30px",
-							mb: "50px",
-						}}
-						size="large"
-						variant="outlined"
-					>
-						Exchange
-					</Button>
-				</Box>
 			</Box>
-		</SupportedAssetsProvider>
+		</Box>
 	);
 };
